Refactor CreateTasksFile func
